backend/requirements.txt
Flask==2.3.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.100.0
